Ensure CLI performance metrics are accurate with SAX infosets
When using the CLI performance subcommand with the SAX infoset type, the
performance loop currently includes serializing SAX events to XML while
parsing and deserialzing XML to SAX events on unparsing. This means that
the performance numbers when using the -I sax option are inflated since
they include serialization/deserialization actions that would normally
be handled outside of Daffodil or potentially never even take place.

This changes the CLI logic performance command, so that on parse it
simply ignores all SAX events and on unparse it converts an XML infoset
to an array of custom SaxEvents outside of the performance loop and then
replays those inside the loop during unparse. This ensures no XML
serialization/deserialiation occurs in the performance loop when using
SAX infosets.

For the CLI parse/unparse commands, we maintain the current behavior for
usability, so that users can see a visualization of the SAX events as
XML, similar to how we pretty print JDOM, Scala XML, etc.

Additionally, the infoset type logic in Main.scala has gotten pretty
complicated, especially with the more recent additions of the SAX API
and EXI support--lots of infoset type logic is spread out and handled
with a scattering of Either objects. To make these changes easier and to
make adding new infoset types easier in the future, all the infoset
handling is refactored out of a number of functions and into separate
InfosetHandler classes. Each such class is specific to an infoset type,
with capabilities split into different functions to separate the logic
into performance loop and non-performance loop capabilities.

DAFFODIL-2400