Changes daffodil-release-candidate to be triggered by a GitHub action
This moves all the logic of building and publishing a release candidate
into a GitHub action. This means that users no longer have to build and
run any containers on their system. The only thing that must be done to
trigger a release trigger the correct workflow providing as input the
release candidate label. Signing keys and dist/nexus authentication
tokens are all provided as GitHub secrets managed by Apache Infra.

To support complete automation, the daffodil-release-candidate script is
changed so that it no longer asks for user input. It now expects all
inputs to be provide as command line options.

Although this is intended to be run via GitHub actions, the README
provides the commands to build and run locally for testing purposes.
The commands do not provide the --publish option and so performs a dry
run that does everything except publish the built artifacts.

Also did some cleanup, like remove repetitive "DAFFODIL_" in variable
names and fix all errors/warnings reported by shellcheck.

DAFFODIL-XXX