Remove bit position buffer used with regex scanning
The regexMatchBitPositionBuffer keeps track of the bit positions of each
decoded character. The purpose of this is so taht when we find which
character ends a regex match we can quickly find the associated bit
position of the character and set our position to that.

Some downsides to this it adds complexity, it requires a LongBuffer the
same as as the CharBuffer, and maybe most significantly it requires a
slow decode algorithm that decodes one character at a time as it records
the bit position of each character.

We can actually remove this position buffer without much loss in
performance in most cases. Most encodings are fixed-width so instead we
can just calculate how many bits the regex matched and add that to the
starting decode position. This doesn't work with non-fixed-width
encodings like UTF-8, so we need to redecode the characters with a
limited char buffer. This adds some overhead, but since this now allows
for faster decoders that don't decode one character at a time, and
doesn't require the LongBuffer, it should hopefully make up for it.

DAFFODIL-XXX