Remove bit position buffer used with delimiter scanning
The bitPositiions buffer keeps track of the bit positions of each
decoded character used in the CharIterator used for delimiter scanning.
The purpose of this is so that when the iterator calls next() we can
return the character and quickly update the bit position to where that
character ended.

Some downsides to this it adds complexity, it requires a LongBuffer the
same as as the CharBuffer, and maybe most significantly it requires a
slow decode algorithm that decodes one character at a time as it records
the bit position of each character.

We can actually remove this position buffer without much loss in
performance in most cases. Most encodings are fixed-width, so when
next() is called we can instead just calculate how many bits a single
character takes and increment the bit position by that amount.

This doesn't work with non-fixed-width encodings like UTF-8, so we need
somethign else. I have no idea what that is, since that also a number of
issues.

DAFFODIL-XXX