Improve character decode performance
Daffodil currently implements custom charset decoders, even for those
with Java implemenations. Not only does this mean we have to have custom
and potentially unoptimized implemenations for complex decoder logic,
they also rely on our default BitsCharsetDecoder which decodes one
characer at a time. This is very slow but was required to support the
bitPositions buffer used to keep track of the each character. All uses
of this bitPosition buffer have been removed, so we are no longer
limited to slow one-char-at-a-time decoding.

For charsets with a Java implementation, we now use a BitsCharsetDecoder
that wraps the Java CharsetDecoder. This allows deleting a lot of custom
decoder logic that Java already implements. It also uses a new
getByteBuffer function added to InputSourceDataInputStream that avoids
copying bytes. Now bytes are decoded in larger chunks, which
can lead to significant decode performance especially as the number of
characters increases.

This also moves all Java-based charset definitions to a single file so
adding new ones doesn't require creating new files.

This also simplifies the BitsCharsetNonByteSizeDecoder that is used by
character sets without a Java implementation. This still reads bytes one
at a time so the decoder doesn't have to worry about unalined data,
fragment bits, bitOrder, etc., but it avoids exception handling and does
all the logic in a single smaller function. These kinds of charsets are
also not that common and when they are used the strings are relativly
infrequent and small, so non-highly optmized performance may be
acceptable for now. So now there are two decoders, one fast one for
charsets with Java implemenations, and a slower one for non charsets
without Java implementations.

DAFFODIL-