Use a custom GitHub action to create release candidates from tags
Although the release candidate container has helped to make builds
reproducible and limit the amount of setup needed, it is still difficult
to setup in some environments. Now that we have reproducible builds, ASF
allows us to automate creating and signing release artifacts in GitHub
Actions.

To implement this, this creates a new custom GitHub action called
"asf-release-candidate" that prepares the GitHub build environment for
creating release candidate for ASF. This includes things like SBT
settings for publishing to ASF maven repositories, creating directories
for artifacts on dist.apache.org, importing signing keys, and creating
the source artifact.Workflows using this action only need ito publish
artifacts (e.g. using sbt publishSigned) and write any helper binaries
to an artifacts directory. The post script of the action will then sign,
checksum, and commit the artifacts.

This also adds a new "Release Candidate" workflow that uses this action
and is triggered when a release candidate tag is created, simplifing the
release candidate process to just pushing a single signed tag. Workflow
dispatch is allowed for testing, but it will not publish anything.

With this change, we no longer need the release candidate container. So
it is renamed and greatly simplfied to just a simple "build release
container". This locally builds artifacts similar to what the new
workflow does, allowing for testing prior to tagging or verifying
reproducible builds after a release candidate is made.

DAFFODIL-XXX