Only use mmap for regular files in the CLI
The CLI uses the result of Files.size() on input files to determine if
it should mmap the file, which has shown to give significant performance
gains. However, for non-regular files, asking for the size or trying to
mmap doesn't make sense, and the behavior is undefined. In practice,
this seems to work, but the result is mapping a buffer of zero bytes,
leading the a CLI parser/unparse seeing no data.

To fix this, the CLI now checks if a file is regular, and only if it is
regular will it consider mmapping the file. All non-regular files use
the existing fallback streaming behavior.

DAFFODIL-3002